+++ rules/aip0121/resource_must_support_list.go
@@ -33,6 +33,12 @@ var resourceMustSupportList = &lint.ServiceRule{
 		// Iterate all RPCs and try to find resources. Mark the
 		// resources which have a List method, and which ones do not.
 		for _, m := range s.GetMethods() {
+			// Streaming methods do not count as standard methods even if they
+			// look like them.
+			if utils.IsStreaming(m) {
+				continue
+			}
+
 			if utils.IsListMethod(m) {
 				if msg := utils.GetListResourceMessage(m); msg != nil && utils.IsResource(msg) {
 					t := utils.GetResource(msg).GetType()